+++ CHANGELOG.md
@@ -6,7 +6,16 @@
 * Support label references to user defined listings environment
 * Add option to disable automatic compilation in power save mode
 * Convert automatic compilation settings to a combobox
+
+* Add experimental support for the addtoluatexpath package
+
+
+
+
 * Add inspection to check for LaTeX package updates
+
+
+
 * Add checkboxes to graphic insertion wizard for relative width or height
 
 ### Fixed